+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.4.0
 
 - Ready launchers are selected first, but then unready launchers are queued up.
+- Bugfix: selecting attack wouldn't set attack command mode first time (no selection)
 
 ## 0.3.0